actualización de contratos

+++ contracts/2ExtraccionMostoContrac.sol
@@ -3,17 +3,17 @@ pragma solidity ^0.8.11;
 
 contract ExtraccionMostoContrac {
     struct ExtraccionMosto {
-        uint id;
+        uint256 id;
         string tipo;
         bool aprobado;
-        uint createdAt;
+        uint256 createdAt;
     }
 
-    mapping(address => mapping(uint => ExtraccionMosto)) public lista;
-    mapping(address => uint) public contador;
+    mapping(address => mapping(uint256 => ExtraccionMosto)) public lista;
+    mapping(address => uint256) public contador;
 
     function crear(string memory _tipo) public {
-        uint contador_id = contador[msg.sender];
+        uint256 contador_id = contador[msg.sender];
 
         lista[msg.sender][contador_id] = ExtraccionMosto(
             contador_id,
@@ -24,7 +24,7 @@ contract ExtraccionMostoContrac {
         contador[msg.sender]++;
     }
 
-    function actualizar(uint _id, string memory _tipo) public {
+    function actualizar(uint256 _id, string memory _tipo) public {
         ExtraccionMosto memory _item = lista[msg.sender][_id];
         if (_item.aprobado == false) {
             _item.tipo = _tipo;
@@ -32,7 +32,7 @@ contract ExtraccionMostoContrac {
         }
     }
 
-    function aprobarProceso(uint _id) public {
+    function aprobarProceso(uint256 _id) public {
         ExtraccionMosto memory _item = lista[msg.sender][_id];
         _item.aprobado = true;
         lista[msg.sender][_id] = _item;
