+++ contracts/3PasteurizacionMostoContrac.sol
@@ -3,23 +3,23 @@ pragma solidity ^0.8.11;
 
 contract PasteurizacionMostoContrac {
     struct Pasteurizacion {
-        uint id;
+        uint256 id;
         string temperatura_caliente;
         string temperatura_fria;
         string tiempo_proceso;
         bool aprobado;
-        uint createdAt;
+        uint256 createdAt;
     }
-    
-    mapping(address => mapping(uint => Pasteurizacion)) public lista;
-    mapping(address => uint) public contador;
+
+    mapping(address => mapping(uint256 => Pasteurizacion)) public lista;
+    mapping(address => uint256) public contador;
 
     function crear(
         string memory _temperatura_caliente,
         string memory _temperatura_fria,
         string memory _tiempo_proceso
     ) public {
-        uint contador_id = contador[msg.sender];
+        uint256 contador_id = contador[msg.sender];
 
         lista[msg.sender][contador_id] = Pasteurizacion(
             contador_id,
@@ -33,7 +33,7 @@ contract PasteurizacionMostoContrac {
     }
 
     function actualizar(
-        uint _id,
+        uint256 _id,
         string memory _temperatura_caliente,
         string memory _temperatura_fria,
         string memory _tiempo_proceso
@@ -47,7 +47,7 @@ contract PasteurizacionMostoContrac {
         }
     }
 
-    function aprobarProceso(uint _id) public {
+    function aprobarProceso(uint256 _id) public {
         Pasteurizacion memory _item = lista[msg.sender][_id];
         _item.aprobado = true;
         lista[msg.sender][_id] = _item;
