+++ contracts/6TrasiegoContrac.sol
@@ -3,18 +3,18 @@ pragma solidity ^0.8.11;
 
 contract TrasiegoContrac {
     struct Trasiego {
-        uint id;
-        uint liquido_claro;
-        uint liquido_oscuro;
+        uint256 id;
+        uint256 liquido_claro;
+        uint256 liquido_oscuro;
         bool aprobado;
-        uint createdAt;
+        uint256 createdAt;
     }
 
-    mapping(address => mapping(uint => Trasiego)) public lista;
-    mapping(address => uint) public contador;
+    mapping(address => mapping(uint256 => Trasiego)) public lista;
+    mapping(address => uint256) public contador;
 
-    function crear(uint _liquido_claro, uint _liquido_oscuro) public {
-        uint contador_id = contador[msg.sender];
+    function crear(uint256 _liquido_claro, uint256 _liquido_oscuro) public {
+        uint256 contador_id = contador[msg.sender];
 
         lista[msg.sender][contador_id] = Trasiego(
             contador_id,
@@ -27,9 +27,9 @@ contract TrasiegoContrac {
     }
 
     function actualizar(
-        uint _id,
-        uint _liquido_claro,
-        uint _liquido_oscuro
+        uint256 _id,
+        uint256 _liquido_claro,
+        uint256 _liquido_oscuro
     ) public {
         Trasiego memory _item = lista[msg.sender][_id];
         if (_item.aprobado == false) {
@@ -39,7 +39,7 @@ contract TrasiegoContrac {
         }
     }
 
-    function aprobarProceso(uint _id) public {
+    function aprobarProceso(uint256 _id) public {
         Trasiego memory _item = lista[msg.sender][_id];
         _item.aprobado = true;
         lista[msg.sender][_id] = _item;
